+++ src/main/java/domain/player/state/Stay.java
@@ -1,14 +1,14 @@
-package domain.state;
+package domain.player.state;
 
 import domain.card.Hand;
 
-public class BlackJack extends Finished {
-    BlackJack(Hand hand) {
+public class Stay extends Finished {
+    public Stay(Hand hand) {
         super(hand);
     }
     
     @Override
     public double calculateProfit(double betAmount) {
-        return betAmount * 1.5;
+        return betAmount;
     }
 }
