+++ src/test/java/domain/player/state/HitTest.java
@@ -1,9 +1,12 @@
-package domain.state;
+package domain.player.state;
 
 import domain.card.Card;
 import domain.card.Hand;
 import domain.card.Number;
 import domain.card.Shape;
+import domain.player.state.Bust;
+import domain.player.state.Hit;
+import domain.player.state.State;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
